chore(headless): share request parameters
