+++ packages/headless/src/api/search/facet-search/facet-search-request.ts
@@ -1,6 +1,6 @@
 import {SpecificFacetSearchRequest} from './specific-facet-search/specific-facet-search-request';
 import {CategoryFacetSearchRequest} from './category-facet-search/category-facet-search-request';
-import {BaseParam} from '../search-api-params';
+import {BaseParam} from '../../platform-service-params';
 
 export type FacetSearchRequest = BaseParam &
   (SpecificFacetSearchRequest | CategoryFacetSearchRequest);
