+++ packages/headless/src/api/search/search/search-request.ts
@@ -1,22 +1,24 @@
 import {
-  AdvancedQueryParam,
   BaseParam,
-  ConstantQueryParam,
   ContextParam,
+  DebugParam,
+  LocaleParam,
+  NumberOfResultsParam,
+  VisitorIDParam,
+} from '../../platform-service-params';
+import {
+  AdvancedQueryParam,
+  ConstantQueryParam,
   EnableDidYouMeanParam,
   EnableQuerySyntaxParam,
   FacetOptionsParam,
   FacetsParam,
   FieldsToIncludeParam,
   FirstResultParam,
-  NumberOfResultsParam,
   PipelineParam,
   QueryParam,
   SearchHubParam,
   SortCriteriaParam,
-  VisitorIDParam,
-  DebugParam,
-  LocaleParam,
 } from '../search-api-params';
 
 export type SearchRequest = BaseParam &
